plans page implemented in web app
