+++ web/magic_ai_text/app/(authenticated)/page.tsx
@@ -1,4 +1,4 @@
-import { MarketingPage } from "../Pages/MarketingPage";
+import { MarketingPage } from "../../components/Marketing/MarketingPage";
 
 export default function Page(){
     return <MarketingPage/>
